v1.33_修正积分8问题，单帧图像帧频显示保留2为小数，界面输入卫星编号。

+++ Spectra/UserControls/Ctrl_ImageView.xaml.cs
@@ -329,7 +329,7 @@ namespace Spectra
             DataRow dr = ImageInfo.dtImgInfo.Rows[(int)p.X];
 
             int frm = Convert.ToInt32(dr["FrameId"]);
-            int freq = Convert.ToInt32(dr["Freq"]);
+            double freq = Convert.ToDouble(dr["Freq"]);
             int integral = Convert.ToInt32(dr["Integral"]);
             int startRow = Convert.ToInt32(dr["StartRow"]);
             int gain = Convert.ToInt32(dr["Gain"]);
